requirements.txt
Package                                   Version
----------------------------------------- -------
asgiref                                   3.5.2
Babel                                     2.11.0
Django                                    4.1.3
django-countries                          7.4.2
django-extensions                         3.2.1
django-filter                             22.1
django-model-utils                        4.2.0
django-money                              3.0.0
djangorestframework                       3.14.0
djangorestframework-serializer-extensions 2.0.1
drf-flex-fields                           1.0.0
hashids                                   1.3.1
pip                                       21.3.1
py-moneyed                                2.0
pytz                                      2022.6
setuptools                                60.2.0
sqlparse                                  0.4.3
typing_extensions                         4.4.0
wheel                                     0.37.1
